requirements.txt
nonebot2[fastapi]
nonebot-adapter-onebot
